fusion/requirements.txt
django==1.8.6
django-pagination==1.0.7
django-extensions==1.4.8
feedparser==5.1.3
django-mailer==0.1.0
MySQL-python==1.2.5
django-pagination-bootstrap==1.0.5
psutil==2.2.1
python-dateutil==2.4.2
xlwt==1.0.0